pyproject.toml
[tool.poetry]
name = "meta_system"
version = "0.1.0"
description = ""
authors = []
license = "Apache 2.0"

[[tool.poetry.source]]
url = "https://pypi.org/simple"
name = "public" # pypi name is reserved...
default = true

[tool.poetry.dependencies]
python = "^3.7"
flask = "^1.1.1"
docker = "^4.2.0"
pydash = "^4.7.6"
pebble = "^4.5.0"
pyyaml = "^5.3"
flask-cors = "^3.0.8"
pyjwt = "^1.7.1"
flask-jwt-extended = "^3.24.1"
flask-bcrypt = "^0.7.1"
psutil = "^5.7.0"
flask-pymongo = "^2.3.0"
flask_mail = "^0.9.1"
uritools = "^3.0.0"
pymodm = "^0.4.2"
pandas = "^1.0.3"
sh = "^1.13.1"
pydantic = "^1.6.1"
pymongo = "^3.11.0"
circus = "^0.17.1"
chaussette = "^1.3.0"
meinheld = "^1.0.2"
mkdocs = "^1.1.2"
tqdm = "^4.51.0"
click = "^7.1.2"

[tool.poetry.dev-dependencies]
poetry = "^1.1.1"
sphinx = "^2.4.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
